[md] raid5-ppl: Partial Parity Log write logging implementation
Message-id: <1491767829-48574-4-git-send-email-ncroxon@redhat.com>
Patchwork-id: 176314
O-Subject: [RHEL7.4 PATCH 3/7] raid5-ppl: Partial Parity Log write logging implementation
Bugzilla: 1434872
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

Pulled changes in from upstream based on the commit id below.
Minor differences between upstream and RHEL except for 3 lines.
They are not applicable to upstream kernels, it's caused by a
customization made by Red Hat (struct bio_aux).

@ -146,6 +146,7 @@ struct ppl_io_unit {
        /* inline bio and its biovec for submitting the iounit */
        struct bio bio;
+       struct bio_aux bio_aux;
        struct bio_vec biovec[PPL_IO_INLINE_BVECS];
 };

@ -215,6 +216,8 @@ static struct ppl_io_unit *ppl_new_iounit(struct ppl_log *log,
        INIT_LIST_HEAD(&io->stripe_list);
        atomic_set(&io->pending_stripes, 0);
        bio_init(&io->bio);
+       io->bio.bio_aux = &io->bio_aux;
+       atomic_set(&io->bio.bio_aux->__bi_remaining, 1);
        io->bio.bi_io_vec = io->biovec;
        io->bio.bi_max_vecs = PPL_IO_INLINE_BVECS;

It would be better to use bio_init_aux() here but it's not exported.
-Nigel Croxon

commit 3418d036c81dcb604b7c7c71b209d5890a8418aa
Author: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Date:   Thu Mar 9 09:59:59 2017 +0100

raid5-ppl: Partial Parity Log write logging implementation

Implement the calculation of partial parity for a stripe and PPL write
logging functionality. The description of PPL is added to the
documentation. More details can be found in the comments in raid5-ppl.c.

Attach a page for holding the partial parity data to stripe_head.
Allocate it only if mddev has the MD_HAS_PPL flag set.

Partial parity is the xor of not modified data chunks of a stripe and is
calculated as follows:

- reconstruct-write case:
  xor data from all not updated disks in a stripe

- read-modify-write case:
  xor old data and parity from all updated disks in a stripe

Implement it using the async_tx API and integrate into raid_run_ops().
It must be called when we still have access to old data, so do it when
STRIPE_OP_BIODRAIN is set, but before ops_run_prexor5(). The result is
stored into sh->ppl_page.

Partial parity is not meaningful for full stripe write and is not stored
in the log or used for recovery, so don't attempt to calculate it when
stripe has STRIPE_FULL_WRITE.

Put the PPL metadata structures to md_p.h because userspace tools
(mdadm) will also need to read/write PPL.

Warn about using PPL with enabled disk volatile write-back cache for
now. It can be removed once disk cache flushing before writing PPL is
implemented.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Signed-off-by: Shaohua Li <shli@fb.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>