[md] add sysfs entries for PPL
Message-id: <1491767829-48574-5-git-send-email-ncroxon@redhat.com>
Patchwork-id: 176318
O-Subject: [RHEL7.4 PATCH 4/7] md: add sysfs entries for PPL
Bugzilla: 1434872
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

I cherry-picked from upstream. The only conflict was the doc changes.
-Nigel Croxon

commit 664aed04446c7f653d8acbe2cdf7989f28238524
Author: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Date:   Thu Mar 9 10:00:00 2017 +0100

md: add sysfs entries for PPL

Add 'consistency_policy' attribute for array. It indicates how the array
maintains consistency in case of unexpected shutdown.

Add 'ppl_sector' and 'ppl_size' for rdev, which describe the location
and size of the PPL space on the device. They can't be changed for
active members if the array is started and PPL is enabled, so in the
setter functions only basic checks are performed. More checks are done
in ppl_validate_rdev() when starting the log.

These attributes are writable to allow enabling PPL for external
metadata arrays and (later) to enable/disable PPL for a running array.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Signed-off-by: Shaohua Li <shli@fb.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>