[md] raid5-ppl: load and recover the log
Message-id: <1491767829-48574-6-git-send-email-ncroxon@redhat.com>
Patchwork-id: 176321
O-Subject: [RHEL7.4 PATCH 5/7] raid5-ppl: load and recover the log
Bugzilla: 1434872
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

Pulled changes in from upstream based on the commit id by Artur.
Rhel7 already backport 2953079c692da067aeb6345659875b97378f9b0a,
a simple change is needed for it compile.

- drivers/md/raid5-ppl.c    2017-04-05 15:48:28.000000000 +0800
+ drivers/md/raid5-ppl1.c    2017-04-05 17:00:37.637199964 +0800
          * without any issues, set the array as clean.
          */
         mddev->recovery_cp = MaxSector;
-        set_bit(MD_CHANGE_CLEAN, &mddev->flags);
+        set_bit(MD_SB_CHANGE_CLEAN, &mddev->sb_flags);
     } else if (mddev->pers && ppl_conf->mismatch_count > 0) {
         /* no mismatch allowed when enabling PPL for a running array */
         ret = -EINVAL;

-Nigel Croxon

commit 4536bf9ba2d03404655586b07f8830b6f2106242
Author: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Date:   Thu Mar 9 10:00:01 2017 +0100

raid5-ppl: load and recover the log

Load the log from each disk when starting the array and recover if the
array is dirty.

The initial empty PPL is written by mdadm. When loading the log we
verify the header checksum and signature. For external metadata arrays
the signature is verified in userspace, so here we read it from the
header, verifying only if it matches on all disks, and use it later when
writing PPL.

In addition to the header checksum, each header entry also contains a
checksum of its partial parity data. If the header is valid, recovery is
performed for each entry until an invalid entry is found. If the array
is not degraded and recovery using PPL fully succeeds, there is no need
to resync the array because data and parity will be consistent, so in
this case resync will be disabled.

Due to compatibility with IMSM implementations on other systems, we
can't assume that the recovery data block size is always 4K. Writes
generated by MD raid5 don't have this issue, but when recovering PPL
written in other environments it is possible to have entries with
512-byte sector granularity. The recovery code takes this into account
and also the logical sector size of the underlying drives.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Signed-off-by: Shaohua Li <shli@fb.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>