[tools] virtio_ring: Support DMA APIs
Message-id: <1491816201-42716-5-git-send-email-wexu@redhat.com>
Patchwork-id: 176328
O-Subject: [PATCH 04/18] virtio_ring: Support DMA APIs
Bugzilla: 1283257 1425127
RH-Acked-by: Dave Young <dyoung@redhat.com>
RH-Acked-by: Jason Wang <jasowang@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>

From: Andy Lutomirski <luto@kernel.org>

virtio_ring currently sends the device (usually a hypervisor)
physical addresses of its I/O buffers.  This is okay when DMA
addresses and physical addresses are the same thing, but this isn't
always the case.  For example, this never works on Xen guests, and
it is likely to fail if a physical "virtio" device ever ends up
behind an IOMMU or swiotlb.

The immediate use case for me is to enable virtio on Xen guests.
For that to work, we need vring to support DMA address translation
as well as a corresponding change to virtio_pci or to another
driver.

Signed-off-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
(cherry picked from commit 780bc7903a32edb63be138487fd981694d993610)
Signed-off-by: Wei Xu <wexu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>