[uapi] virtio: new feature to detect IOMMU device quirk
Message-id: <1491816201-42716-10-git-send-email-wexu@redhat.com>
Patchwork-id: 176333
O-Subject: [PATCH 09/18] virtio: new feature to detect IOMMU device quirk
Bugzilla: 1425127 1283257
RH-Acked-by: Dave Young <dyoung@redhat.com>
RH-Acked-by: Jason Wang <jasowang@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>

From: "Michael S. Tsirkin" <mst@redhat.com>

The interaction between virtio and IOMMUs is messy.

On most systems with virtio, physical addresses match bus addresses,
and it doesn't particularly matter which one we use to program
the device.

On some systems, including Xen and any system with a physical device
that speaks virtio behind a physical IOMMU, we must program the IOMMU
for virtio DMA to work at all.

On other systems, including SPARC and PPC64, virtio-pci devices are
enumerated as though they are behind an IOMMU, but the virtio host
ignores the IOMMU, so we must either pretend that the IOMMU isn't
there or somehow map everything as the identity.

Add a feature bit to detect that quirk: VIRTIO_F_IOMMU_PLATFORM.

Any device with this feature bit set to 0 needs a quirk and has to be
passed physical addresses (as opposed to bus addresses) even though
the device is behind an IOMMU.

Note: it has to be a per-device quirk because for example, there could
be a mix of passed-through and virtual virtio devices. As another
example, some devices could be implemented by an out of process
hypervisor backend (in case of qemu vhost, or vhost-user) and so support
for an IOMMU needs to be coded up separately.

It would be cleanest to handle this in IOMMU core code, but that needs
per-device DMA ops. While we are waiting for that to be implemented, use
a work-around in virtio core.

Note: a "noiommu" feature is a quirk - add a wrapper to make
that clear.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
(cherry picked from commit 1a937693993ff10d7e80cca6ddd55f3000aa6376)
Signed-off-by: Wei Xu <wexu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	include/linux/virtio_config.h
	drivers/virtio/virtio_ring.c
virtio_config.h: Cherry-pick is confused by existing comment symbol.
virtio_ring.c: Following upstream commit is dropped in RHEL since we don't
support virtio-on-Xen.

commit 78fe39872378
Author: Andy Lutomirski <luto@kernel.org>
Date:   Tue Feb 2 21:46:40 2016 -0800

    vring: Use the DMA API on Xen