[redhat] spec: add requirement for new version kexec-tools
Message-id: <1492395433-28736-1-git-send-email-piliu@redhat.com>
Patchwork-id: 176764
O-Subject: [rhel7.4 PATCH v2] kernel.spec: add requirement for new version kexec-tools
Bugzilla: 1432299
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Pratyush Anand <panand@redhat.com>
RH-Acked-by: Dave Young <dyoung@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 1432299
https://bugzilla.redhat.com/show_bug.cgi?id=1432299

The kernel for rhel7.4 has merged KASLR-related patches. These patches cause
the old version makedumpfile fails with segment fault. So if a user just
upgrades the kernel from rhel7.3 to 7.4, it brings issue. This patch aims
to force the user to upgrade the kexec-tools with new kernel.

Known issue:
kexec-tools-2.0.14-x requires dracut >= 033-449, which can not be satisfied
in 7.2. Since dracut has dependency of systemd/udev and other packages,
it is hard to upgrade and may fail. But even if users can not update kexec-tools
in 7.2, it is better than one does not notice the conflicts at all.

Signed-off-by: Pingfan Liu <piliu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>