[pci] Avoid output of ioapic probe error
Message-id: <1492624951-8035-1-git-send-email-darcari@redhat.com>
Patchwork-id: 176857
O-Subject: [RESENT RHEL7.4 PATCH BZ 1377960] Avoid output of ioapic probe error
Bugzilla: 1377960
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Bugzilla:
---------
http://bugzilla.redhat.com/1377960

Description:
------------
This patch avoids output of the following dmesg error:

    ioapic: probe of <pci address> failed with error -22

The error is harmlessas as the APICs are loaded at boot time from the MADT.
The ioapic_probe is called later on for devices that are already present and
have no entries in acpi tables.

This has been the subject of a number of bug reports and has resulted in a
kbase article:

    https://access.redhat.com/solutions/913863

The problem is an artifact of the upstream kernel from which RHEL was first
cut.  At that time, the ioapic driver was being modified to work with the
standard device model in the kernel.  Return values were not verified and
as a result an invalid error message is output for each ioapic device. For
example:

    ioapic: probe of 0000:00:05.4 failed with error -22

Upstream removed this driver in favor of a hotplug driver which has never been
required for RHEL7.  Backporting the driver requires significant changes which
will probably never make it to RHEL7:

    https://bugzilla.redhat.com/show_bug.cgi?id=1174199#c3

There is no loss of functionality associated with this error message.  This
simple one line patch avoids the error by returning -ENODEV instead of -EINVAL.
Based on this return value, really_probe() in dd.c does not output the error
message.

Upstream:
---------
RHEL_only

Brew:
-----
https://brewweb.devel.redhat.com/taskinfo?taskID=13022129

Testing:
--------
Booted -650.el7 and observed the following errors:

[    3.652224] ioapic: probe of 0000:00:05.4 failed with error -22
[    3.658833] ioapic: probe of 0000:80:05.4 failed with error -22

Rebooted with brew build and no errors were observed.
--------

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>