[x86] revert "mm: Simplify get_user_pages() PTE bit handling"
Message-id: <20170424172000.17313-3-jmoyer@redhat.com>
Patchwork-id: 177050
O-Subject: [RHEL7.4 PATCH 2/5] revert: [x86] mm: Simplify get_user_pages() PTE bit handling
Bugzilla: 1443284
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

BZ: 1443284
Brew: 13054327

This patch set caused a general protection fault in gup_huge_pud.
This particular patch is not the problem, but it is layered on
top of the problematic patch.  Revert it for now, and then re-apply
after the problematic patch is removed.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>