[net] sched: align nlattr properly when needed
Message-id: <20170504211919.23079-2-ivecera@redhat.com>
Patchwork-id: 177656
O-Subject: [RHEL7.4 net PATCH 1/5] sched: align nlattr properly when needed
Bugzilla: 1442088
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>
RH-Acked-by: Paolo Abeni <pabeni@redhat.com>
RH-Acked-by: Marcelo Leitner <mleitner@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Davide Caratti <dcaratti@redhat.com>

BZ#1442088

Backport note: Actions bpf, connmark & ife are not present in RHEL so
the relevant hunks were stripped. The padattr field was added to struct
gnet_dump into 4-bytes hole instead of original place to avoid kABI
breakage report. This breakage is reported because pointer to this
struct is passed to dump_stats callback of Qdisc_ops and
Qdisc_class_ops. In any case extending the struct by RH_KABI_EXTEND
should be safe as well as the qdisc API is not and should not be whitelisted.

Upstream commit(s):
commit 9854518ea04db33738602d45ebc96a200e6f5198
Author: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Date:   Tue Apr 26 10:06:18 2016 +0200

    sched: align nlattr properly when needed

    Signed-off-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
    Signed-off-by: David S. Miller <davem@davemloft.net>

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>