[vfio] type1: Remove locked page accounting workqueue
Message-id: <20170505210558.16300.31776.stgit@gimli.home>
Patchwork-id: 177716
O-Subject: [RHEL-7.4 kernel PATCH 1/3] vfio/type1: Remove locked page accounting workqueue
Bugzilla: 1438403
RH-Acked-by: Auger Eric <eric.auger@redhat.com>
RH-Acked-by: Peter Xu <peterx@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1438403
Upstream: 0cfef2b7410b64d7a430947e0b533314c4f97153
Notes: No down_write_killable() in RHEL, so we use down_write() and
       eliminate the error path where the killable version could fail.
       This also means that @ret needs explicit initialization.

If the mmap_sem is contented then the vfio type1 IOMMU backend will
defer locked page accounting updates to a workqueue task.  This has a
few problems and depending on which side the user tries to play, they
might be over-penalized for unmaps that haven't yet been accounted or
race the workqueue to enter more mappings than they're allowed.  The
original intent of this workqueue mechanism seems to be focused on
reducing latency through the ioctl, but we cannot do so at the cost
of correctness.  Remove this workqueue mechanism and update the
callers to allow for failure.  We can also now recheck the limit under
write lock to make sure we don't exceed it.

vfio_pin_pages_remote() also now necessarily includes an unwind path
which we can jump to directly if the consecutive page pinning finds
that we're exceeding the user's memory limits.  This avoids the
current lazy approach which does accounting and mapping up to the
fault, only to return an error on the next iteration to unwind the
entire vfio_dma.

Cc: stable@vger.kernel.org
Reviewed-by: Peter Xu <peterx@redhat.com>
Reviewed-by: Kirti Wankhede <kwankhede@nvidia.com>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>