[fs] nfs: use safe, interruptible sleeps when waiting to retry LOCK
Message-id: <055c0cf50481a05a73c617cb62eee8c6f6d6d104.1494249737.git.bcodding@redhat.com>
Patchwork-id: 177887
O-Subject: [RHEL7 fs PATCH v3 05/23] nfs: use safe, interruptible sleeps when waiting to retry LOCK
Bugzilla: 1377710
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

We actually want to use TASK_INTERRUPTIBLE sleeps when we're in the
process of polling for a NFSv4 lock. If there is a signal pending when
the task wakes up, then we'll be returning an error anyway. So, we might
as well wake up immediately for non-fatal signals as well. That allows
us to return to userland more quickly in that case, but won't change the
error that userland sees.

Also, there is no need to use the *_unsafe sleep variants here, as no
vfs-layer locks should be held at this point.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Anna Schumaker <Anna.Schumaker@Netapp.com>
(cherry picked from commit 66f570ab7307ad613b0d800fff27aba316014eaf)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>