[fs] nfs: allow blocking locks to be awoken by lock callbacks
Message-id: <b68ff5182e6ea665002c71602fc6b3bcf4221b38.1494249737.git.bcodding@redhat.com>
Patchwork-id: 177890
O-Subject: [RHEL7 fs PATCH v3 11/23] nfs: allow blocking locks to be awoken by lock callbacks
Bugzilla: 1377710
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Add a waitqueue head to the client structure. Have clients set a wait
on that queue prior to requesting a lock from the server. If the lock
is blocked, then we can use that to wait for wakeups.

Note that we do need to do this "manually" since we need to set the
wait on the waitqueue prior to requesting the lock, but requesting a
lock can involve activities that can block.

However, only do that for NFSv4.1 locks, either by compiling out
all of the waitqueue handling when CONFIG_NFS_V4_1 is disabled, or
skipping all of it at runtime if we're dealing with v4.0, or v4.1
servers that don't send lock callbacks.

Note too that even when we expect to get a lock callback, RFC5661
section 20.11.4 is pretty clear that we still need to poll for them,
so we do still sleep on a timeout. We do however always poll at the
longest interval in that case.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
[Anna: nfs4_retry_setlk() "status" should default to -ERESTARTSYS]
Signed-off-by: Anna Schumaker <Anna.Schumaker@Netapp.com>

(cherry picked from commit a1d617d8f134679741b0b35e8e1436b015ac5538)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>