[fs] nfsd: add a LRU list for blocked locks
Message-id: <9a8700dbc9c61f672f0717ff023537bd1c34773b.1494249738.git.bcodding@redhat.com>
Patchwork-id: 177882
O-Subject: [RHEL7 fs PATCH v3 21/23] nfsd: add a LRU list for blocked locks
Bugzilla: 1377710
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

It's possible for a client to call in on a lock that is blocked for a
long time, but discontinue polling for it. A malicious client could
even set a lock on a file, and then spam the server with failing lock
requests from different lockowners that pile up in a DoS attack.

Add the blocked lock structures to a per-net namespace LRU when hashing
them, and timestamp them. If the lock request is not revisited after a
lease period, we'll drop it under the assumption that the client is no
longer interested.

This also gives us a mechanism to clean up these objects at server
shutdown time as well.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
(cherry picked from commit 7919d0a27f1e7cb324e023776aa1cbff00f1ee7b)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>