[fs] gfs2: Allow glocks to be unlocked after withdraw
Message-id: <10693811.5362489.1494259025974.JavaMail.zimbra@redhat.com>
Patchwork-id: 177937
O-Subject: [RHEL7.4 fs PATCH] GFS2: Allow glocks to be unlocked after withdraw
Bugzilla: 1404005
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Hi,

This patch fixes a regression introduced by patch 0d1c7ae9d8.

The intent of the patch was to stop promoting glocks after a
file system is withdrawn due to a variety of errors, because doing
so results in a BUG(). (You should be able to unmount after a
withdraw rather than having the kernel panic.)

Unfortunately, it also stopped demotions, so glocks could not be
unlocked after withdraw, which means the unmount would hang.

This patch allows function do_xmote to demote locks to an
unlocked state after a withdraw, but not promote them.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

This is a RHEL7 port of the following upstream patch:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/fs/gfs2?id=ed17545d01e4a40086f1622c88ce64cc0241cfc2

A kernel RHEL7.4 build was done here:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=13151048

It was tested on system gfs-i16c-01.mpc.lab.eng.bos.redhat.com
with a reliable reproducer documented in the bz, which
faithfully recreates the regression.

This is a necessary addendum to the following bug:

rhbz#1404005
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>