[redhat] makefile.cross: check and install kernel.spec build requirements
Message-id: <1495040083-27592-1-git-send-email-prarit@redhat.com>
Patchwork-id: 178481
O-Subject: [RHEL7.4 INTERNAL] cross: check and install kernel.spec build requirements
Bugzilla:
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

jbenc noticed a 'make rh-cross-build-all' failure in which the x86_64
build failed because the java openjdk-devel package was not installed.
The java-devel requirement was added to the kernel.spec for perf and is a
Provides for java openjdk-devel.

The RPMs listed in Makefile.cross are rpms that are required for the cross
to complete and do not take into account the build requirements for the
kernel itself.  These requirements are available as BuildRequires in the
kernel.spec and the cross target(s) should check to see if they are
installed.

rh-cross targets should check and install the kernel.spec build
requirements.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Cc: jbenc@redhat.com
Cc: Rafael Aquini <aquini@redhat.com>
Cc: dzickus@redhat.com
Signed-off-by: Rafael Aquini <aquini@redhat.com>