[drm] i915: don't rcu-sync from shrinker
Message-id: <20170519211648.11728-1-rclark@redhat.com>
Patchwork-id: 178752
O-Subject: [RHEL7 drm] drm/i915: don't rcu-sync from shrinker
Bugzilla: 1443028
RH-Acked-by: Dave Airlie <airlied@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>

Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=13242258
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1443028
Upstream: n/a (downstream hack for crusty old kernel)

This is a bit sad, but on RHEL7 kernel it is not safe to synchronize
with RCU grace period from shrinker.  This has the effect that i915's
request kmem_cache could transiently consume a lot of memory if
userspace somehow managed to generate a lot of execbuf ioctls within
the RCU grace period.

Upstream, what happens is the shrinker throttles on RCU under i915's
struct_mutex, blocking incoming execbuf's under memory pressure.
But we can't do that in the RHEL7 kernel without backporting the
following RCU commits (and all they depend on):

  6587a23b - rcu: Switch synchronize_sched_expedited() to IPI
  807226e2 - rcu: Stop excluding CPU hotplug in synchronize_sched_expedited()

Which would be a rather significant change, especially after the end
of devel phase.

So instead we live with the (mostly theoretical) issue, to avoid:

   =================================
   [ INFO: inconsistent lock state ]
   3.10.0-664.el7.x86_64.debug #1 Tainted: G           OE  ------------ T
   ---------------------------------
   inconsistent {RECLAIM_FS-ON-W} -> {IN-RECLAIM_FS-R} usage.
   kswapd0/43 [HC0[0]:SC0[0]:HE1:SE1] takes:
    (cpu_hotplug.lock){+++++?}, at: [<ffffffffa909c7e7>] get_online_cpus+0x37/0x80
   {RECLAIM_FS-ON-W} state was registered at:
     [<ffffffffa911f8a9>] mark_held_locks+0xb9/0x140
     [<ffffffffa912277d>] lockdep_trace_alloc+0x6d/0xd0
     [<ffffffffa922973a>] kmem_cache_alloc_node_trace+0x3a/0x360
     [<ffffffffa90d60d5>] __smpboot_create_thread.part.1+0x35/0xf0
     [<ffffffffa90d66ed>] smpboot_create_threads+0x6d/0xa0
     [<ffffffffa909ca97>] _cpu_up+0x97/0x1d0
     [<ffffffffa909cc81>] cpu_up+0xb1/0x130
     [<ffffffffa9e6c523>] smp_init+0x95/0x97
     [<ffffffffa9e431ae>] kernel_init_freeable+0xf1/0x217
     [<ffffffffa9759e1e>] kernel_init+0xe/0xf0
     [<ffffffffa97826d8>] ret_from_fork+0x58/0x90
   irq event stamp: 59
   hardirqs last  enabled at (59): [<ffffffffa9777a66>] _raw_spin_unlock_irqrestore+0x36/0x70
   hardirqs last disabled at (58): [<ffffffffa97782eb>] _raw_spin_lock_irqsave+0x2b/0xa0
   softirqs last  enabled at (0): [<ffffffffa9096154>] copy_process+0x794/0x1c90
   softirqs last disabled at (0): [<          (null)>]           (null)

   other info that might help us debug this:
    Possible unsafe locking scenario:

          CPU0
          ----
     lock(cpu_hotplug.lock);
     <Interrupt>
       lock(cpu_hotplug.lock);

    *** DEADLOCK ***

   2 locks held by kswapd0/43:
    #0:  (shrinker_rwsem){++++..}, at: [<ffffffffa91d92ac>] shrink_slab+0x3c/0x520
    #1:  (&dev->struct_mutex){+.+.+.}, at: [<ffffffffc0449f93>] i915_gem_shrinker_lock+0x33/0x70 [i915]

   stack backtrace:
   CPU: 0 PID: 43 Comm: kswapd0 Tainted: G           OE  ------------ T 3.10.0-664.el7.x86_64.debug #1
   Hardware name: Intel To be filled by O.E.M./To be filled by O.E.M., BIOS 4.6.3 04/28/2010
    ffffa0a5aade8000 00000000e82b7c23 ffffa0a5aadef8f0 ffffffffa976d3a5
    ffffa0a5aadef940 ffffffffa976764c 0000000000000000 ffffa0a500000001
    ffffa0a500000001 000000000000000a ffffa0a5aade8000 ffffffffa911e3a0
   Call Trace:
    [<ffffffffa976d3a5>] dump_stack+0x19/0x1b
    [<ffffffffa976764c>] print_usage_bug+0x1f7/0x208
    [<ffffffffa911e3a0>] ? check_usage_backwards+0x180/0x180
    [<ffffffffa911f76d>] mark_lock+0x21d/0x2a0
    [<ffffffffa91205ed>] __lock_acquire+0x76d/0x15f0
    [<ffffffffa9121d59>] lock_acquire+0x99/0x1e0
    [<ffffffffa909c7e7>] ? get_online_cpus+0x37/0x80
    [<ffffffffa909c80d>] get_online_cpus+0x5d/0x80
    [<ffffffffa909c7e7>] ? get_online_cpus+0x37/0x80
    [<ffffffffa916f884>] synchronize_sched_expedited+0x64/0x1f0
    [<ffffffffa916fbfe>] synchronize_rcu_expedited+0xe/0x10
    [<ffffffffc044a520>] i915_gem_shrink+0x310/0x4d0 [i915]
    [<ffffffffc0449f93>] ? i915_gem_shrinker_lock+0x33/0x70 [i915]
    [<ffffffffc044a8c1>] i915_gem_shrinker_scan+0x71/0xb0 [i915]
    [<ffffffffc024506e>] shrinker2_shrink+0x1e/0x40 [drm]
    [<ffffffffa91d9329>] shrink_slab+0xb9/0x520
    [<ffffffffa923e605>] ? mem_cgroup_iter+0xc5/0x3e0
    [<ffffffffa91ddd29>] balance_pgdat+0x4d9/0x610
    [<ffffffffa91de01b>] kswapd+0x1bb/0x660
    [<ffffffffa90cc530>] ? wake_up_atomic_t+0x30/0x30
    [<ffffffffa91dde60>] ? balance_pgdat+0x610/0x610
    [<ffffffffa90cb1ed>] kthread+0xed/0x100
    [<ffffffffa90cb100>] ? insert_kthread_work+0x80/0x80
    [<ffffffffa97826d8>] ret_from_fork+0x58/0x90
    [<ffffffffa90cb100>] ? insert_kthread_work+0x80/0x80

Signed-off-by: Rob Clark <rclark@redhat.com>
---
See upstream 0eafec6d3244802d469712682b0f513963c23eff for background.

I would welcome feedback about possibility of RCU backport for rhel75.
I think it is a bit late in the game for this for rhel74, and I'm not
sure offhand what kabi challenges might be lurking.  But maybe this
would also be useful for other subsystems?

I guess since I am bringing up the subject, I am volunteering for the
backport, unless we have some RCU expert in-house who wants to do this.
If we *did* go this route for rhel75, I guess it would have to be
pretty early in the cycle, and I think I would need some help for
testing this on non-desktop workloads.

OTOH I think I'm ok with living with this limitation and in the off-
chance that it results in any bz's because of the workaround in i915,
moving them all to rhel8.

Signed-off-by: Rafael Aquini <aquini@redhat.com>