[nvme] nvme-fabrics: Allow ctrl loss timeout configuration
Message-id: <1495573291-18208-3-git-send-email-dmilburn@redhat.com>
Patchwork-id: 178876
O-Subject: [RHEL7.4 PATCH 2/3] nvme-fabrics: Allow ctrl loss timeout configuration
Bugzilla: 1452321 1442735
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

From: Sagi Grimberg <sagi@grimberg.me>

https://bugzilla.redhat.com/show_bug.cgi?id=1452321
https://bugzilla.redhat.com/show_bug.cgi?id=1442735

When a host sense that its controller session is damaged,
it tries to re-establish it periodically (reconnect every
reconnect_delay). It may very well be that the controller
is gone and never coming back, in this case the host will
try to reconnect forever.

Add a ctrl_loss_tmo to bound the number of reconnect attempts
to a specific controller (default to a reasonable 10 minutes).
The timeout configuration is actually translated into number of
reconnect attempts and not a schedule on its own but rather
divided with reconnect_delay. This is useful to prevent
racing flows of remove and reconnect, and it doesn't really
matter if we remove slightly sooner than what the user requested.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Sagi Grimberg <sagi@grimberg.me>
Signed-off-by: Jens Axboe <axboe@fb.com>
(cherry picked from commit 42a45274c225829637b585e1c994a68f59507108)
Signed-off-by: David Milburn <dmilburn@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>